Change license to BSL 1.0

+++ include/data_tree/common/common.hpp
@@ -2,32 +2,6 @@
  * @brief Common utilities
  * @author Matthew Guidry (github: mguid65)
  * @date 2024-02-05
- *
- * @cond IGNORE_LICENSE
- *
- * MIT License
- *
- * Copyright (c) 2024 Matthew Guidry
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *
- * @endcond
  */
 
 #ifndef DATATREE_COMMON_HPP
@@ -270,7 +244,7 @@ namespace key_literals {
  * @param idx an index
  * @return a KeyType created from an index
  */
-inline KeyOrIdxType operator""_k(unsigned long long idx) { return KeyOrIdxType{idx}; }
+inline KeyOrIdxType operator""_k(unsigned long long idx) { return KeyOrIdxType{static_cast<std::size_t>(idx)}; }
 
 /**
  * @brief Helper for string literal KeyType UDL
